refactor: rename animation to Vertical Parallax Gallery
