+++ frontend/src/App.jsx
@@ -12,7 +12,7 @@ import SandboxWrapperTest from "./components/SandboxWrapper/SandboxWrapperTest";
 import Animation from "./components/Animation/Animation";
 import ZoomAnimation from "./components/ZoomAnimation/ZoomAnimation";
 import HorizontalGalleryDemo from "./components/HorizontalGallery/HorizontalGalleryDemo";
-import ParallaxGalleryPage from "./components/ParallaxGallery/ParallaxPage";
+import VerticalParallaxGalleryPage from "./components/VerticalParallaxGallery/VerticalParallaxPage";
 import WipePage from "./components/WipeGallery/WipePage";
 
 function App() {
@@ -39,7 +39,7 @@ function App() {
       <Route path="/zoom-anim" element={<ZoomAnimation />} />
       <Route path="/animation" element={<Animation />} />
       <Route path="/horizontal-gallery" element={<HorizontalGalleryDemo />} />
-      <Route path="/parallax-gallery" element={<ParallaxGalleryPage />} />
+      <Route path="/vertical-parallax-gallery" element={<VerticalParallaxGalleryPage />} />
       <Route path="/wipe-gallery" element={<WipePage />} />
     </Routes>
   )
