+++ frontend/src/components/Nav.jsx
@@ -20,7 +20,7 @@ function Nav() {
           <Link to="/text" onClick={() => setOpen(false)}>Text</Link>
           <Link to="/forms" onClick={() => setOpen(false)}>Forms</Link>
           <Link to="/horizontal-gallery" onClick={() => setOpen(false)}>Horizontal Gallery</Link>
-          <Link to="/parallax-gallery" onClick={() => setOpen(false)}>Parallax Gallery</Link>
+          <Link to="/vertical-parallax-gallery" onClick={() => setOpen(false)}>Vertical Parallax Gallery</Link>
           <Link to="/wipe-gallery" onClick={() => setOpen(false)}>Wipe Gallery</Link>
         </div>
       )}
